backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.12
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
statsmodels==0.14.0
xgboost==2.0.3
scipy==1.11.4
python-multipart==0.0.6
python-dotenv==1.0.0
